// src/app/components/item/item.interface.ts
export interface IItem {
  itemCode: string;
  itemName: string;
  itemType: string;
  itemStatus?: string;
  itemPrice: string;
  manufactureCountry?: string;
  manufacturerItemDescription?: string;
  allowDiscount?: string;
  qtyInPackage: string;
  quantity?: string;
  unitOfMeasure?: string;
  unitOfMeasurePrice?: string;
  unitQty?: string;
  manufacturerName?: string;
  bIsWeighted?: string;
  priceUpdateDate?: string;
  image?: string;
}